+++ src/App.js
@@ -1,86 +1,20 @@
 import './App.css';
-import { useState, useEffect } from 'react'
 import SearchAppBar from './components/searchBar';
-import TitlebarBelowImageList from './components/image-list';
+import { Home } from './pages/Home'
+import { Search } from './pages/SearchPage'
+import { Gallery } from './pages/PersonalGallery'
+import { Routes, Route } from "react-router-dom";
 function App() {
-  // const [listImages, setListImages] = useState(null);
-  useEffect(() => {
-    // IMPROVE to https://api.unsplash.com/photos/?client_id={process.env.REACT_APP_CLIENT_ID}
-    fetch('./mocks/list-images-gallery.json')
-      .then(response => response.json())
-      .then(data => {
-        // setListImages(data)
-        console.table(data)
-      })
-  }, [])
   return (
     <div className="App">
       <SearchAppBar></SearchAppBar>
-      <TitlebarBelowImageList itemData={itemData} />
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/search" element={<Search />} />
+        <Route path="/gallery" element={<Gallery />} />
+      </Routes>
     </div>
   );
 }
 
 export default App;
-const itemData = [
-  {
-    img: 'https://images.unsplash.com/photo-1551963831-b3b1ca40c98e',
-    title: 'Breakfast',
-    author: '@bkristastucchio',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1551782450-a2132b4ba21d',
-    title: 'Burger',
-    author: '@rollelflex_graphy726',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1522770179533-24471fcdba45',
-    title: 'Camera',
-    author: '@helloimnik',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1444418776041-9c7e33cc5a9c',
-    title: 'Coffee',
-    author: '@nolanissac',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1533827432537-70133748f5c8',
-    title: 'Hats',
-    author: '@hjrc33',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1558642452-9d2a7deb7f62',
-    title: 'Honey',
-    author: '@arwinneil',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1516802273409-68526ee1bdd6',
-    title: 'Basketball',
-    author: '@tjdragotta',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1518756131217-31eb79b20e8f',
-    title: 'Fern',
-    author: '@katie_wasserman',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1597645587822-e99fa5d45d25',
-    title: 'Mushrooms',
-    author: '@silverdalex',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1567306301408-9b74779a11af',
-    title: 'Tomato basil',
-    author: '@shelleypauls',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1471357674240-e1a485acb3e1',
-    title: 'Sea star',
-    author: '@peterlaster',
-  },
-  {
-    img: 'https://images.unsplash.com/photo-1589118949245-7d38baf380d6',
-    title: 'Bike',
-    author: '@southside_customs',
-  },
-];
