+++ src/pages/Home.js
@@ -1,0 +1,10 @@
+export const Home = () => {
+  return (
+    <>
+      <main>
+        <h2>Welcome to the homepage!</h2>
+        <p>You can do this, I believe in you.</p>
+      </main>
+    </>
+  );
+}
