+++ src/pages/SearchPage.js
@@ -1,0 +1,13 @@
+export const Search = () => {
+  return (
+    <>
+      <main>
+        <h2>Who are we?</h2>
+        <p>
+          That feels like an existential question, don't you
+          think?
+        </p>
+      </main>
+    </>
+  );
+}
